make helper functions
