+++ v1/get_capy_hour.go
@@ -4,10 +4,7 @@ import (
 	"fmt"
 	"image"
 	"io/ioutil"
-	"strconvert"
-	"math/rand"
 	"os"
-	"time"
 
 	"github.com/gofiber/fiber/v2"
 	"github.com/looskie/capybara-api/utils"
@@ -17,24 +14,11 @@ func GetCapyHour(c *fiber.Ctx) error {
 	var wantsJSON = utils.WantsJSON(c)
 
 	// sets seed for this hour
-	var date = time.Now()
-	var hour = date.Hour()
-	var day = date.Day()
-	var month = strconv.Itoa(int(date.Month()))
-	var year = strconv.Itoa(date.Year())
-	var joined = year + month + day + hour
-	seed, err := strconv.Atoi(joined)
-	if err != nil {
-		println(err.Error())
-	}
-	rand.Seed(int64(seed))
+	utils.SetSeed("hour")
 
-	// find boundary of largest number rand should generate
-	files,_ := ioutil.ReadDir("capys/")
-    var max_rand = len(files)
+	// get random index
+	var index = utils.GetIndex()
 
-	// set index
-	var index = rand.Intn(max_rand) + 1
 	bytes, err := ioutil.ReadFile("capys/capy" + fmt.Sprint(index) + ".jpg")
 
 	c.Set("X-Capybara-Index", fmt.Sprint(index))
